SCMBOB-671 fix overwriting file
The `AddFiles` method of NuGet will only write a file if it doesn't exist
yet. We tried to fix this by deleting every single file of NuGet package
before installing the package. We did this with the `RemoveFiles` of
NuGet. However this method is extermely sensible. Even if there is a
read-handle on a file, it won't delete it.
To finaly fix the issue, we iterate ourself over the file collection and
use the `AddFile` method, which overwrites the file. One could expect that
this is the same as AddFiles, but the AddFiles method call the AddFile
method if a file doesn't exist.